# Improve RGB Min Max evaluation performance by using 2 or 3 comparison…