Fixes: https://github.com/dotnet/runtime/issues/46153

…s instead of 4

* Replaces 2 calls each to Math.Min and Math.Max (a total of 4 required
  comparisons) with a single call to MinMaxRgb (at least 2 required
  comparisons and at worst 3).