@L2 2021-04-01
dotnet/performance microbenchmarks filtered with System.Drawing.Tests.Perf_Color*:

Base = .NET6 Before this changeset
Diff = .NET6 After this changeset

No Slower results for the provided threshold = 2% and noise filter = 25ns.

| Faster                                        | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| --------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Drawing.Tests.Perf_Color.GetHue        |      1.23 |          1756.53 |          1432.47 |         |
| System.Drawing.Tests.Perf_Color.GetBrightness |      1.09 |          1544.74 |          1415.17 |         |
| System.Drawing.Tests.Perf_Color.GetSaturation |      1.09 |          1602.51 |          1476.70 |         |