@L2 2021-04-02
> Do we have sufficient functional test coverage?

Thanks Dan. There were a couple of test cases missing which I've just added:

All Possible Cases:
| label for existing test | case |
| ----- | ----:|
| (1) | r	==	g	==	b |
| (2) | r	==	g	<	b |	
| (3) | r	==	g	>	b |	
| (4) | r	<	g	==	b |	
| (missing) | r	<	g	<	b |	
| | r	<	g	>	b |
| (5) | where [r == b] |				
| (missing) | where [r > b]  |					
| (missing) | where [r < b] |					
| (7) | r	>	g	==	b |	
| | r	>	g	<	b |
| (6) | where [r == b]  |				
| (missing) | where [r > b] |					
| (missing) | where [r < b] |					
| (missing) | r	>	g	>	b |	


Existing Test Coverage:
| test label | case |
| ----- | ----:|
| (1) | [InlineData(0, 0, 0, 0f)] |
|  | [InlineData(255, 255, 255, 1f)] |
| (7) | [InlineData(255, 0, 0, 0.5f)] |		
| (5) | [InlineData(0, 255, 0, 0.5f)] |		
| (2) | [InlineData(0, 0, 255, 0.5f)] |		
| (3) | [InlineData(255, 255, 0, 0.5f)] |		
| (6) | [InlineData(255, 0, 255, 0.5f)] |		
| (4) | [InlineData(0, 255, 255, 0.5f)] |		
|  | [InlineData(51, 255, 255, 0.6f)] |
|  | [InlineData(255, 51, 255, 0.6f)] |
|  | [InlineData(255, 255, 51, 0.6f)] |
|  | [InlineData(255, 51, 51, 0.6f)] |
|  | [InlineData(51, 255, 51, 0.6f)] |
|  | [InlineData(51, 51, 255, 0.6f)] |
|  | [InlineData(51, 51, 51, 0.2f)] |

For:
```
public void GetBrightness(int r, int g, int b, float expected)
{
    Assert.Equal(expected, Color.FromArgb(r, g, b).GetBrightness());
}
```