@L2 2021-04-02
Also I wasn't sure if I should put this new method in Math.cs and make it a general use case to find the min and max of three values so I've left it inside Color.cs for now.