@danmoseley 2021-04-02
> wasn't sure if I should put this new method in Math.cs

Drawing wouldn't be able to use anything from Math unless it was public: the libraries can only use each other's public surface. And to make new public API involves a review process, which would include consideration of how broadly useful the API would be. So I think it is in the right place.