@L2 2021-04-02
> Other than small style improvements, looks good. Thanks!

Thanks @safern , I've updated it with the requested changes.