@L2 2021-04-06
> Thanks for the contribution @L2 - perhaps you'd like to pick up something else?

No problem :) , and sure thing @danmoseley , I'd love to help improve performance in any way I can. I have a few more changes I'm looking into that are in the pipeline and hopefully many more to come as part of my work at AMD.